--- a/.travis.yml
+++ b/.travis.yml
@@ -154,7 +154,7 @@ env:
     <<: *cron-only
     <<: *make-test-only
     env:
-      - BEFORE_INSTALL="sed vm_opts.h -e 's/SUPPORT_JOKE *0/SUPPORT_JOKE 1/' -i"
+      - BEFORE_INSTALL="sed vm_opts.h -e 's/OPT_SUPPORT_JOKE *0/OPT_SUPPORT_JOKE 1/' -i"
 
   - &WITH_COROUTINE_UCONTEXT
     name: COROUTINE=ucontext