Defining unused instructions increases the code size and might be cache-inefficient.

I also think it would be less confusing if `RubyVM::INSTRUCTION_NAMES` were not including never-used instruction names.