@k0kubun 2019-09-02
Actually there seems to be no significant impact for the code size and the performance:

```
$ du -s .before/ruby
19992   .before/ruby
$ du -s .after/ruby
19968   .after/ruby
```

```
$ benchmark-driver -v benchmark.yml --rbenv 'before;after' --repeat-count=24 --output=all
before: ruby 2.7.0dev (2019-09-02T16:53:49Z master a848b62819) [x86_64-linux]
after: ruby 2.7.0dev (2019-09-02T17:51:48Z define-insn-if 20c1eba480) [x86_64-linux]
Calculating -------------------------------------
                                       before                 after
Optcarrot Lan_Master.nes    34.59277689908502     44.77167767437890 fps
                            50.78644504597501     50.22970798490342
                            50.87373949035694     50.97830020418768
                            50.96296563885826     51.19192144088704
                            51.23114971773986     51.25107782165509
                            51.30573112913191     51.39678004625837
                            51.60580817079894     51.43262754954327
                            51.76309561454278     51.52897399857301
                            51.85526051352603     51.69479225111350
                            51.92934807599517     51.77463040990341
                            51.99210824939702     51.82493654649841
                            52.03118052436776     52.01979002642318
                            52.08272541401557     52.06308114653079
                            52.10750663221185     52.18035556723671
                            52.15555460322726     52.23022935037871
                            52.24166913375453     52.23201382634296
                            52.27232136532778     52.24100108353972
                            52.37950154530666     52.24127216357662
                            52.43442734171256     52.36145752523225
                            52.54790912856974     52.39189827581511
                            52.62162637851512     52.42357449420416
                            52.70071551326700     52.49261411073728
                            52.72387067831448     52.49775264515822
                            52.79516801831293     52.98736776605598
```

Nevertheless, I think this behavior is more natural and nice to have to avoid confusion.