@methodmissing 2019-09-02
Nice work. Also more specifically to see layout changes in the text segment, eg. :

```
lourens@CarbonX1:~/src/ruby/trunk$ size ./ruby
   text	   data	    bss	    dec	    hex	filename
3807713	  22120	  73280	3903113	 3b8e89	./ruby
```

Also for icache impact on Linux with perf tool install, `sudo perf stat -ddd -f benchmark-driver -v benchmark.yml --rbenv 'before;after' --repeat-count=24 --output=all`. Didn't know about the `output=all` flag, neat.

Output example (optcarrot runs long enough for a decent and representative sample size):

```
       4264,662878      task-clock (msec)         #    0,913 CPUs utilized          
              2442      context-switches          #    0,573 K/sec                  
              2370      cpu-migrations            #    0,556 K/sec                  
             36678      page-faults               #    0,009 M/sec                  
        7571576961      cycles                    #    1,775 GHz                      (29,89%)
       10775147326      instructions              #    1,42  insn per cycle           (37,83%)
        2250460899      branches                  #  527,700 M/sec                    (38,18%)
          52418348      branch-misses             #    2,33% of all branches          (38,33%)
        2778000925      L1-dcache-loads           #  651,400 M/sec                    (38,39%)
         108385922      L1-dcache-load-misses     #    3,90% of all L1-dcache hits    (38,76%)
          32742208      LLC-loads                 #    7,678 M/sec                    (31,05%)
           4095596      LLC-load-misses           #   12,51% of all LL-cache hits     (31,15%)
   <not supported>      L1-icache-loads                                             
         181129258      L1-icache-load-misses                                         (31,14%) <<<<<<<<<<
        2980154756      dTLB-loads                #  698,802 M/sec                    (31,35%)
           4699184      dTLB-load-misses          #    0,16% of all dTLB cache hits   (30,77%)
           2802029      iTLB-loads                #    0,657 M/sec                    (30,52%) <<<<<<<<<<
           1371393      iTLB-load-misses          #   48,94% of all iTLB cache hits   (30,46%) <<<<<<<<
   <not supported>      L1-dcache-prefetches                                        
   <not supported>      L1-dcache-prefetch-misses  
```

Relevant is `L1-dcache-loads`,  `L1-icache-load-misses`, `iTLB-loads` and `iTLB-load-misses`